requirements.txt
fastapi
uvicorn
python-telegram-bot
motor
google-genai
python-dotenv
